Fix building on FreeBSD
Closes https://github.com/owncloud/ocis/issues/1217

* Extracted common error methods to `errtypes` package
* Fixed type error on Bavail ( https://golang.org/src/syscall/ztypes_freebsd_arm64.go#L125 )
* Alias ENODATA to ENOATTR

How to test this on Linux:

```
export GOOS=freebsd
make
```
