+++ Makefile
@@ -1,6 +1,6 @@
 .PHONY: build
 
-SHELL := /bin/bash
+SHELL := bash
 BUILD_DATE=`date +%FT%T%z`
 GIT_COMMIT=`git rev-parse --short HEAD`
 GIT_BRANCH=`git rev-parse --symbolic-full-name --abbrev-ref HEAD`
