+++ changelog/unreleased/freebsd-build.md
@@ -1,0 +1,5 @@
+Bugfix: Fix build on FreeBSD
+
+FreeBSD doesn't define ENODATA and Statfs.Bavail is int64
+
+https://github.com/cs3org/reva/pull/1499
