+++ pkg/storage/fs/owncloudsql/owncloudsql_freebsd.go
@@ -16,9 +16,9 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-// +build !windows,!freebsd
+// +build freebsd
 
-package owncloud
+package owncloudsql
 
 import (
 	"context"
@@ -63,11 +63,11 @@ func calcEtag(ctx context.Context, fi os.FileInfo) string {
 	if err != nil {
 		log.Error().Err(err).Msg("error writing size")
 	}
-	etag := fmt.Sprintf(`"%x"`, h.Sum(nil))
-	return fmt.Sprintf("\"%s\"", strings.Trim(etag, "\""))
+	etag := fmt.Sprintf("%x", h.Sum(nil))
+	return strings.Trim(etag, "\"")
 }
 
-func (fs *ocfs) GetQuota(ctx context.Context) (uint64, uint64, error) {
+func (fs *owncloudsqlfs) GetQuota(ctx context.Context) (uint64, uint64, error) {
 	// TODO quota of which storage space?
 	// we could use the logged in user, but when a user has access to multiple storages this falls short
 	// for now return quota of root
@@ -76,7 +76,7 @@ func (fs *ocfs) GetQuota(ctx context.Context) (uint64, uint64, error) {
 	if err != nil {
 		return 0, 0, err
 	}
-	total := stat.Blocks * uint64(stat.Bsize)                // Total data blocks in filesystem
-	used := (stat.Blocks - stat.Bavail) * uint64(stat.Bsize) // Free blocks available to unprivileged user
+	total := stat.Blocks * uint64(stat.Bsize)                        // Total data blocks in filesystem
+	used := (stat.Blocks - uint64(stat.Bavail)) * uint64(stat.Bsize) // Free blocks available to unprivileged user
 	return total, used, nil
 }
