+++ pkg/storage/fs/owncloudsql/owncloudsql_unix.go
@@ -16,7 +16,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-// +build !windows
+// +build !windows,!freebsd
 
 package owncloudsql
