+++ pkg/storage/utils/decomposedfs/node/node.go
@@ -157,9 +157,9 @@ func ReadNode(ctx context.Context, lu PathLookup, id string) (n *Node, err error
 	switch {
 	case err == nil:
 		n.ParentID = string(attrBytes)
-	case isNoData(err):
+	case errtypes.XattrIsNoData(err):
 		return nil, errtypes.InternalError(err.Error())
-	case isNotFound(err):
+	case errtypes.XattrIsNotFound(err):
 		return n, nil // swallow not found, the node defaults to exists = false
 	default:
 		return nil, errtypes.InternalError(err.Error())
@@ -187,7 +187,7 @@ func ReadNode(ctx context.Context, lu PathLookup, id string) (n *Node, err error
 	// Check if parent exists. Otherwise this node is part of a deleted subtree
 	_, err = os.Stat(lu.InternalPath(n.ParentID))
 	if err != nil {
-		if isNotFound(err) {
+		if errtypes.XattrIsNotFound(err) {
 			return nil, errtypes.NotFound(err.Error())
 		}
 		return nil, err
@@ -290,7 +290,7 @@ func (n *Node) Owner() (*userpb.UserId, error) {
 	switch {
 	case err == nil:
 		owner.OpaqueId = string(attrBytes)
-	case isNoData(err), isNotFound(err):
+	case errtypes.XattrIsNoData(err), errtypes.XattrIsNotFound(err):
 		fallthrough
 	default:
 		return nil, err
@@ -301,7 +301,7 @@ func (n *Node) Owner() (*userpb.UserId, error) {
 	switch {
 	case err == nil:
 		owner.Idp = string(attrBytes)
-	case isNoData(err), isNotFound(err):
+	case errtypes.XattrIsNoData(err), errtypes.XattrIsNotFound(err):
 		fallthrough
 	default:
 		return nil, err
@@ -312,7 +312,7 @@ func (n *Node) Owner() (*userpb.UserId, error) {
 	switch {
 	case err == nil:
 		owner.Type = utils.UserTypeMap(string(attrBytes))
-	case isNoData(err), isNotFound(err):
+	case errtypes.XattrIsNoData(err), errtypes.XattrIsNotFound(err):
 		fallthrough
 	default:
 		// TODO the user type defaults to invalid, which is the case
@@ -632,9 +632,9 @@ func readChecksumIntoResourceChecksum(ctx context.Context, nodePath, algo string
 			Type: storageprovider.PKG2GRPCXS(algo),
 			Sum:  hex.EncodeToString(v),
 		}
-	case isNoData(err):
+	case errtypes.XattrIsNoData(err):
 		appctx.GetLogger(ctx).Debug().Err(err).Str("nodepath", nodePath).Str("algorithm", algo).Msg("checksum not set")
-	case isNotFound(err):
+	case errtypes.XattrIsNotFound(err):
 		appctx.GetLogger(ctx).Error().Err(err).Str("nodepath", nodePath).Str("algorithm", algo).Msg("file not fount")
 	default:
 		appctx.GetLogger(ctx).Error().Err(err).Str("nodepath", nodePath).Str("algorithm", algo).Msg("could not read checksum")
@@ -654,9 +654,9 @@ func readChecksumIntoOpaque(ctx context.Context, nodePath, algo string, ri *prov
 			Decoder: "plain",
 			Value:   []byte(hex.EncodeToString(v)),
 		}
-	case isNoData(err):
+	case errtypes.XattrIsNoData(err):
 		appctx.GetLogger(ctx).Debug().Err(err).Str("nodepath", nodePath).Str("algorithm", algo).Msg("checksum not set")
-	case isNotFound(err):
+	case errtypes.XattrIsNotFound(err):
 		appctx.GetLogger(ctx).Error().Err(err).Str("nodepath", nodePath).Str("algorithm", algo).Msg("file not fount")
 	default:
 		appctx.GetLogger(ctx).Error().Err(err).Str("nodepath", nodePath).Str("algorithm", algo).Msg("could not read checksum")
@@ -686,9 +686,9 @@ func readQuotaIntoOpaque(ctx context.Context, nodePath string, ri *provider.Reso
 		} else {
 			appctx.GetLogger(ctx).Error().Err(err).Str("nodepath", nodePath).Str("quota", string(v)).Msg("malformed quota")
 		}
-	case isNoData(err):
+	case errtypes.XattrIsNoData(err):
 		appctx.GetLogger(ctx).Debug().Err(err).Str("nodepath", nodePath).Msg("quota not set")
-	case isNotFound(err):
+	case errtypes.XattrIsNotFound(err):
 		appctx.GetLogger(ctx).Error().Err(err).Str("nodepath", nodePath).Msg("file not found when reading quota")
 	default:
 		appctx.GetLogger(ctx).Error().Err(err).Str("nodepath", nodePath).Msg("could not read quota")
@@ -813,7 +813,7 @@ func (n *Node) ReadUserPermissions(ctx context.Context, u *userpb.User) (ap prov
 		switch {
 		case err == nil:
 			AddPermissions(&ap, g.GetPermissions())
-		case isNoData(err):
+		case errtypes.XattrIsNoData(err):
 			err = nil
 			appctx.GetLogger(ctx).Error().Interface("node", n).Str("grant", grantees[i]).Interface("grantees", grantees).Msg("grant vanished from node after listing")
 			// continue with next segment
